+++ pages/index.js
@@ -1,68 +1,204 @@
-import Head from "next/head";
-import styles from "../styles/Home.module.css";
+import React, { useState } from "react";
+import "./../styles/UserHomePage.module.css";
+import logo from "./../public/beeholder-logo.png";
+const InputText = (props) => {
+  return (
+    <>
+      <input
+        className="inputText"
+        placeholder={props.placeholder}
+        value={props.value}
+        onChange={props.handleChange}
+      />
+    </>
+  );
+};
+
+const InputSubmit = (props) => {
+  const [input, setInput] = useState("");
+  const handleChange = (event) => {
+    setInput(event.target.value);
+  };
+  return (
+    <>
+      <div className={"InputSubmit"}>
+        <InputText
+          value={input}
+          placeholder={props.placeholder}
+          handleChange={handleChange}
+        />
+        <button
+          className={"button buttonfixedSize"}
+          onClick={() => {
+            props.handleClick(input);
+          }}
+        >
+          {props.buttonText}
+        </button>
+      </div>
+    </>
+  );
+};
+const ImageHolder = (props) => {
+  return (
+    <>
+      <div className={"image-holder"}>
+        <img src={"/beeholder-logo.png"} className={"image " + props.size} />
+      </div>
+    </>
+  );
+};
+const VTTName = () => {
+  return (
+    <>
+      <div className={"VTT-name"}>
+        <ImageHolder size="medium" />
+        <h3 className={"h3"}>Beer Holder VTT</h3>
+      </div>
+    </>
+  );
+};
+const UserProfileAcess = () => {
+  return (
+    <>
+      <div className={"profile-acess-div  flex-align-center"}>
+        <ImageHolder size="small" />
+        <button
+          className={"button buttonFit"}
+          onClick={() => {
+            console.log("hehe");
+          }}
+        >
+          User Options
+        </button>
+        <button
+          className={"button buttonFit"}
+          onClick={() => {
+            console.log("hehe");
+          }}
+        >
+          Compendium
+        </button>
+      </div>
+    </>
+  );
+};
+
+const PjSheetListHolder = (props) => {
+  return (
+    <>
+      <div className={"flex-column"}>
+        <h3 className={"h3 flex-align-center"}>Suas Fichas</h3>
+        <PJSheetList pjs={["estes", "estes", "estes", "estes"]} />
+      </div>
+    </>
+  );
+};
+
+const PJSheet = (props) => {
+  return (
+    <>
+      <div className={"VTT-name"}>
+        <ImageHolder size="small" />
+        <h2>{props.name}</h2>
+      </div>
+    </>
+  );
+};
+const PJSheetList = (props) => {
+  return (
+    <>
+      <ul className={"ul"}>
+        {props.pjs.map((pj) => {
+          return (
+            <li>
+              <PJSheet name={pj} />
+            </li>
+          );
+        })}
+      </ul>
+    </>
+  );
+};
+const Header = () => {
+  return (
+    <>
+      <div className="header">
+        <VTTName />
+        <SessionAcess />
+      </div>
+    </>
+  );
+};
+
+const SessionAcess = () => {
+  return (
+    <>
+      <div className={"SessionAcess"}>
+        <InputSubmit
+          handleClick={(value) => {
+            console.log(value);
+          }}
+          placeholder="Nome da sala"
+          buttonText={"Criar sala"}
+        />
+        <InputSubmit
+          handleClick={(value) => {
+            console.log(value);
+          }}
+          placeholder="ID da sala"
+          buttonText={"Acessar Sala"}
+        />
+      </div>
+    </>
+  );
+};
+
+const SessionList = (props) => {
+  const sessions = [
+    "estes",
+    "estes",
+    "estes",
+    "estes",
+    "estes",
+    "estes",
+    "estes",
+  ];
+  return (
+    <div className={"sessions-grid"}>
+      {sessions.map((session) => {
+        return <SessionItem name={session} />;
+      })}
+    </div>
+  );
+};
+
+const SessionItem = (props) => {
+  return (
+    <>
+      <div className={"flex-column flex-align-center"}>
+        <div id={"session-name"}>{props.name}</div>
+        <ImageHolder size={"big"} />
+      </div>
+    </>
+  );
+};
 
 export default function Home() {
   return (
-    <div className={styles.container}>
-      <Head>
-        <title>Create Next App</title>
-        <meta name="description" content="Generated by create next app" />
-        <link rel="icon" href="/favicon.ico" />
-      </Head>
-
-      <main className={styles.main}>
-        <h1 className={styles.title}>
-          Welcome to <a href="https://nextjs.org">Next.js!</a>
-        </h1>
-
-        <p className={styles.description}>
-          Get started by editing{" "}
-          <code className={styles.code}>pages/index.js</code>
-        </p>
-
-        <div className={styles.grid}>
-          <a href="https://nextjs.org/docs" className={styles.card}>
-            <h2>Documentation &rarr;</h2>
-            <p>Find in-depth information about Next.js features and API.</p>
-          </a>
-
-          <a href="https://nextjs.org/learn" className={styles.card}>
-            <h2>Learn &rarr;</h2>
-            <p>Learn about Next.js in an interactive course with quizzes!</p>
-          </a>
-
-          <a
-            href="https://github.com/vercel/next.js/tree/master/examples"
-            className={styles.card}
-          >
-            <h2>Examples &rarr;</h2>
-            <p>Discover and deploy boilerplate example Next.js projects.</p>
-          </a>
-
-          <a
-            href="https://vercel.com/new?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
-            className={styles.card}
-          >
-            <h2>Deploy &rarr;</h2>
-            <p>
-              Instantly deploy your Next.js site to a public URL with Vercel.
-            </p>
-          </a>
-        </div>
-      </main>
-
-      <footer className={styles.footer}>
-        <a
-          href="https://vercel.com?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Powered by{" "}
-          <span className={styles.logo}>
-            <img src="/vercel.svg" alt="Vercel Logo" width={72} height={16} />
-          </span>
-        </a>
-      </footer>
+    <div className={"root"}>
+      <div>
+        <Header />
+      </div>
+      <div>
+        <UserProfileAcess />
+      </div>
+      <div>
+        <SessionList />
+      </div>
+      <div>
+        <PjSheetListHolder />
+      </div>
     </div>
   );
 }
